requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
fastapi==0.111.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
lxml==5.4.0
mcp==1.9.4
outcome==1.3.0.post0
pillow==11.2.1
pydantic==2.11.6
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==2.3.6
starlette==0.37.2
trio==0.30.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
XlsxWriter==3.2.3
